+++ LinkUs.Modules.RemoteShell/Commands/KillShell.cs
@@ -3,6 +3,6 @@ namespace LinkUs.Modules.RemoteShell.Commands
     public class KillShell : ShellMessage
     {
         public KillShell() { }
-        public KillShell(double processId) : base(processId) { }
+        public KillShell(int processId) : base(processId) { }
     }
 }
