+++ LinkUs.Modules.RemoteShell/Events/ShellEnded.cs
@@ -2,10 +2,10 @@ namespace LinkUs.Modules.RemoteShell.Events
 {
     public class ShellEnded : ShellMessage
     {
-        public double ExitCode { get; set; }
+        public int ExitCode { get; set; }
 
         public ShellEnded() { }
-        public ShellEnded(double exitCode, double processId) : base(processId)
+        public ShellEnded(int exitCode, int processId) : base(processId)
         {
             ExitCode = exitCode;
         }
