+++ LinkUs.Modules.RemoteShell/Events/ShellStarted.cs
@@ -3,6 +3,6 @@ namespace LinkUs.Modules.RemoteShell.Events
     public class ShellStarted : ShellMessage
     {
         public ShellStarted() { }
-        public ShellStarted(double processId) : base(processId) { }
+        public ShellStarted(int processId) : base(processId) { }
     }
 }
